Rename HapConfig to HapConfigTenSQR
Much of the changes in MiscFunctions.HapConfig have to do with
making it compatible with TenSQR. For now, it is easier to simply
maintain two versions of the class and making the two classes distinct.
May be prudent to revisit this in the future.

Also, some minor formatting changes in Orig2Recons3.java.

In HapConfigTenSQR.java, trailing whitespaces are trimmed and all tab
characters converted to 4 spaces.