0. Renamed Main.java to Entrance.java
1. The user command-line input "prefix" has been refactored to a field
in GenParameter as "project_name". It will only be used as prefix of
global files (not individual pools). User command-line input of
"num_pools" is also removed. Now the only user command-line input is the
"property file"

2. Names of each pool will be automatically learned from the list of
files in the folder; and double-checked if another folder of similar
structure will be read in the same function.

3. Many files in intermediate folder have been split into multiple
sub-folders that will be created during the running time.

4. File names "XXX_inter_freq_vars.txt" have been renamed to
"XXX_inter_freq_haps.txt" to reflect the right contents. This applies
for both local and global haplotypes.

5. Pool-IDs will be the names of the samples. So no assignment of
pool-IDs using indexes in the code at all. The initial IDs are set by
the SAM files.

6. Hap-IDs will be in the format of "h7" instead of just the index.
Here, "h" indicate it is a haplotype; the number, e.g., 7 is the 16-base
(hex)integer of the 0/1 string (7= 00111 in this example)

7. Changed several names of local indexes to make them self explanatory.

8. Added more comments to explain variables

9. removed main() in BAMFormatterGATK and PairedReadLinker.

10. TODO1: No change to BAMFormatterAlt has been made yet. In the
future, we need to combine the it with BAMFormatterGATK

11. TODO2: It doesn't work if AEM fails.

Signed-off-by: Quan Long <quanlong1978@gmail.com>