Multiply a constant to avoid numerical singularity of SVD decomposition. This further reduced AEM failing. But this won’t impact the mathematical property at all, as all Importance Factors will be normalized afterwards. When linking regions, most often we have singular matrix or too small numbers: now we  no longer
By introducing level-3 regions, and increasing the number of iterations,
now the error rate is almost zero for a region of 30 SNPs (before
linking) or 60 SNPs (after linking). The perfect local haps ensure
perfect linking! (In the past, the error rate is about 4-5 errors for a
region of 30 SNPs or 8-10 for 60 SNPs.)

When linking two level I regions, a num_mismatch_cufoff paramaeter was
used. If number of mismatches bases between two level I region and the
level II region is less than the cutoff, 2^N regional haplotyoes are
constructed (N is the number of mismathces).